refactor(common): migrate password reset request/response types to shared common package
- Create packages/common/src/requests/ folder with password-reset.request.ts
- Create packages/common/src/responses/ folder with password-reset.response.ts
- Add index exports for requests and responses in common package
- Create main index.ts in common package for centralized exports
- Update common package.json with request/response exports
- Update auth.controller.ts to reference common types in JSDoc
- Remove duplicate password-reset.types.ts from backend/src/types/
- Establish pattern: all controller request/response types in @yacc/common
- This enables frontend and backend to share the same types
