--- packages/backend/src/types/password-reset.types.ts
@@ -1,22 +1,0 @@
-/**
- * Password Reset Types
- * Types for password reset functionality
- */
-
-export interface ForgotPasswordRequest {
-  email: string;
-}
-
-export interface ForgotPasswordResponse {
-  message: string;
-}
-
-export interface ResetPasswordRequest {
-  token: string;
-  newPassword: string;
-}
-
-export interface ResetPasswordResponse {
-  success: boolean;
-  message: string;
-}
