+++ packages/common/package.json
@@ -4,9 +4,14 @@
   "private": true,
   "type": "module",
   "exports": {
+    ".": "./dist/index.js",
     "./types/*": "./dist/types/*.js",
     "./schemas/*": "./dist/schemas/*.js",
-    "./constants/*": "./dist/constants/*.js"
+    "./constants/*": "./dist/constants/*.js",
+    "./requests": "./dist/requests/index.js",
+    "./requests/*": "./dist/requests/*.js",
+    "./responses": "./dist/responses/index.js",
+    "./responses/*": "./dist/responses/*.js"
   },
   "scripts": {
     "build": "tsc",
