+++ packages/common/src/index.ts
@@ -1,0 +1,20 @@
+/**
+ * @yacc/common - Shared types, schemas, requests, and responses
+ * 
+ * This package contains all shared types between frontend and backend
+ */
+
+// Re-export requests
+export type { ForgotPasswordRequest, ResetPasswordRequest } from './requests';
+
+// Re-export responses
+export type { ForgotPasswordResponse, ResetPasswordResponse } from './responses';
+
+// Re-export schemas (if needed by both frontend and backend)
+// export { ForgotPasswordSchema, ResetPasswordSchema } from './schemas';
+
+// Re-export constants
+// export { ... } from './constants';
+
+// Re-export types
+// export type { ... } from './types';
