+++ packages/common/src/requests/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Request DTOs
+ * Centralized exports for all request types
+ */
+
+export type { ForgotPasswordRequest, ResetPasswordRequest } from './password-reset.request';
