+++ packages/common/src/requests/password-reset.request.ts
@@ -1,0 +1,13 @@
+/**
+ * Password Reset Request Types
+ * Shared request DTOs for password reset endpoints
+ */
+
+export interface ForgotPasswordRequest {
+  email: string;
+}
+
+export interface ResetPasswordRequest {
+  token: string;
+  newPassword: string;
+}
