+++ packages/common/src/responses/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Response DTOs
+ * Centralized exports for all response types
+ */
+
+export type { ForgotPasswordResponse, ResetPasswordResponse } from './password-reset.response';
