+++ packages/common/src/responses/password-reset.response.ts
@@ -1,0 +1,13 @@
+/**
+ * Password Reset Response Types
+ * Shared response DTOs for password reset endpoints
+ */
+
+export interface ForgotPasswordResponse {
+  message: string;
+}
+
+export interface ResetPasswordResponse {
+  success: boolean;
+  message: string;
+}
